api-src/requirements.txt
amqp==5.0.6
aniso8601==7.0.0
asgiref==3.3.4
billiard==3.6.4.0
bleach==3.3.0
celery==5.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.2
django-celery==3.1.17
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-jazzmin==2.4.7
django-js-asset==1.2.2
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
kombu==5.1.0
Markdown==3.3.4
packaging==20.9
Pillow==8.2.0
promise==2.3
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pyparsing==2.4.7
pytz==2021.1
Rx==1.6.1
shortuuid==1.0.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
